AMBARI-13595. Stop-and-Start Upgrade: Handle Storm in all upgrade paths (alejandro)

+++ ambari-server/src/main/resources/common-services/STORM/0.9.1.2.1/package/scripts/service.py
@@ -18,6 +18,8 @@ limitations under the License.
 
 """
 
+import os
+
 from resource_management.core.resources import Execute
 from resource_management.core.resources import File
 from resource_management.core.shell import as_user
@@ -76,14 +78,15 @@ def service(name, action = 'start'):
 
   elif action == "stop":
     process_dont_exist = format("! ({no_op_test})")
-    pid = get_user_call_output.get_user_call_output(format("! test -f {pid_file} ||  cat {pid_file}"), user=params.storm_user)[1]
+    if os.path.exists(pid_file):
+      pid = get_user_call_output.get_user_call_output(format("! test -f {pid_file} ||  cat {pid_file}"), user=params.storm_user)[1]
 
-    Execute(format("{sudo} kill {pid}"),
-      not_if = process_dont_exist)
+      Execute(format("{sudo} kill {pid}"),
+        not_if = process_dont_exist)
 
-    Execute(format("{sudo} kill -9 {pid}"),
-      not_if = format(
-        "sleep 2; {process_dont_exist} || sleep 20; {process_dont_exist}"),
-      ignore_failures = True)
+      Execute(format("{sudo} kill -9 {pid}"),
+        not_if = format(
+          "sleep 2; {process_dont_exist} || sleep 20; {process_dont_exist}"),
+        ignore_failures = True)
 
-    File(pid_file, action = "delete")
+      File(pid_file, action = "delete")
